src/requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-appconf==1.0.4
django-filter==2.3.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
